Initial commit for the application skeleton project.
At the moment this is more a demonstration of the autoconf tool chain and a
useful debug "echo" tool for driving /dev/rddma by hand.

It should be grown into a full fledged skeleton sample app using option
parsing libraries, config file saving/loading, etc. as well as an
example of linking and using the api project libraries.

But we gotta start somewhere...