Add a wait option to the sync_find command.
If you want to wait in a frmwrk script for some remote sync object use

	sync_find://name.location?wait

the processing loop will loop indefinitely until the object is found.
This is a cheap proof of concept for the frmwrk application scripts.
Later if we see the need we can add

	sync_find://name.location?wait(<num_attempts>)

and at some point I want to make the timout on fabric remote calls, the
only place where the timeouts are applied at the moment, configurable on
any function call with

	<some_op>://.....?timeout(n)

If timeout isn't given on the call use the default set in the driver. This is
currently hard coded but should be a location parameter

	location_create://name.location?default_timeout(n)